Fix 'Internal Server Error' while trying to delete non existing part

+++ src/main/java/com/formulaoneapi/service/PartService.java
@@ -29,6 +29,9 @@ public class PartService {
     }
 
     public void remove(int id) {
+        if (!partRepository.existsById(id)) {
+            throw new NoSuchElementException(String.format("Part with id '%d' not found", id));
+        }
         partRepository.deleteById(id);
     }
